+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
-- No unreleased changes.
+### Added
+
+- Added a test for the bookmarklet version of the library to ensure it works correctly.
 
 ## [1.2.12] - 2025-05-26